+++ docs/changelog.md
@@ -5,7 +5,8 @@ Forrest is constantly evolving. Sometimes so fast that you might not even get ar
 - **develop**
   - Optional parameter - We introduced the optional parameter that marks a commands parameter as not required. This can be combined with the pre- and suffix option.
   - Some new constraints for parameter validation: url, ip address, mac address and email address. [read more](formats/yaml-format.md#constraints--optional-)
-
+  - We are collecting some data on the commands that are run. Only the command identifier - only if it comes from our API - and the status (success, failure) will be transferred. 
+  
 
 - **2.2.0**
   - Enum can ask for custom values. The user is able to combine enums with custom value by using the `enum-allow-custom` parameter field. [read more](formats/yaml-format.md#command-fields)